* src/main.cc (restore_packages_flow): Check if there's a 	packages.backup file nuder /var/lib/hildon-application-manager 	before trying to restore a backup (N83585).
	* hildon-application-manager.sh: Check if file pointed by $pkgs
	exists before trying to copy it to /home/user (N83585).



git-svn-id: https://garage.maemo.org/svn/hildon-app-mgr/trunk@1170 e55152ac-3a23-0410-980c-f88739b87167

+++ hildon-application-manager.sh
@@ -27,6 +27,8 @@ if [ ! -d $HOME/.hildon-application-manager ]; then
   mkdir $HOME/.hildon-application-manager
 fi
 
-cp "$pkgs" "$upkgs"
+if [ -f $pkgs ]; then
+  cp "$pkgs" "$upkgs"
+fi
 
 exit 0
